+++ service/src/main/kotlin/fi/espoo/vekkuli/service/BoatReservationService.kt
@@ -15,6 +15,7 @@ import fi.espoo.vekkuli.repository.filter.boatspacereservation.LocationExpr
 import fi.espoo.vekkuli.utils.*
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
+import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.MonthDay
@@ -79,9 +80,9 @@ data class ReserveBoatSpaceInput(
     val reservationId: Int,
     val boatId: Int?,
     val boatType: BoatType,
-    val width: Double,
-    val length: Double,
-    val depth: Double,
+    val width: BigDecimal,
+    val length: BigDecimal,
+    val depth: BigDecimal,
     val weight: Int?,
     val boatRegistrationNumber: String?,
     val boatName: String?,
